k8s_jmeter/fast_api/requirements.txt
fastapi
uvicorn
pydantic
httpx
requests
pyjwt